research-agent/requirements.txt
# Backend
arxiv==2.1.0
requests==2.31.0
pydantic==2.5.3
python-dotenv==1.0.0
schedule==1.2.1
openai==1.12.0

# API
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Frontend
streamlit==1.31.0
